archive-image-search/pyproject.toml
[project]
name = "archive-image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.49",
    "chromadb>=1.2.0",
    "pillow>=12.0.0",
    "s3fs>=2025.9.0",
    "st-files-connection>=0.1.0",
    "standard-imghdr>=3.13.0",
    "streamlit>=1.22.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
